moved build workfile to 'openpype.pipeline.workfile'
